Remove BookingView tests and add AvailableRooList tests. Fix AvailableRoomList booking buttons' collapse after booking.
